+++ pkg/reporting/pagination.go
@@ -2,5 +2,9 @@ package reporting
 
 // Pagination limits the number of results for a Report.
 type Pagination struct {
-	// TODO
+	// Offset limits the number of results. Offset <= 0 means no offset.
+	Offset int
+
+	// Limit limits the number of results. Limit <= 0 means unlimited.
+	Limit int
 }